upgrade the readme
