+++ README.md
@@ -35,25 +35,38 @@ Each `Mnemonic` has access to at least one of `Functions`. `Function`s specify t
 
 ## Development
 
-1. Create `.env` file in the project root with the following variables:
+1. create `api-client` user for a given env
+    You can use the [`create-api-user`](https://github.com/QuanTemplate/scripts/tree/master/create-api-user)
 
+2. Create `.env` file in the project root with the following variables:
+
+    ```
+    CAPITALIQ_API_USERNAME=<api username you got with Capital IQ API license>
+    CAPITALIQ_API_PASSWORD=<password for the corresponding Capital IQ API user>
+    QT_ENV=<name of the Quantemplate environment, should be `prod` for any consumers>
+    QT_AUTH_REALM=<name of the Quantemplate auth ream, should be `qt` for any consumers>
+    QT_CLIENT_ID=<id of the api-client user generated by the Quantemplate team>
+    QT_CLIENT_SECRET=＜password for the used api-client user>
+    ```
+
+    for example, for the QT dev env:
     ```
     CAPITALIQ_API_USERNAME=apiadmin@quantemplate.com
     CAPITALIQ_API_PASSWORD=<password from dev credentials>
     QT_ENV=dev
+    QT_AUTH_REALM=test
+    QT_CLIENT_ID=<id of the api-client user>
+    QT_CLIENT_SECRET=＜password for the used api-client user>
     ```
+
     and export it:
 
     ```sh
     export $(xargs < .env)
     ```
 
-2. run the script
+3. run the script
 
     ```sh
-    sbt "capitaliq/run <absolute path to excel file>"
-    ```
-    example:
-    ```sh
-    sbt "capitaliq/run /Users/gbielski/Projects/qt/data-ingress/CapitalIQ.xlsx"
+    sbt "capitaliq/run"
     ```
