--- todo.md
@@ -1,4 +1,0 @@
-1. upload excel file to qt
-  - [] generate api user through script on dev https://github.com/QuanTemplate/scripts/tree/master/create-api-user
-  - [] get token from the keycloak https://quantemplate.readme.io/docs/setting-up-a-connection
-  - [] send xlsx file to qt with the token
